refactor(api) remove bot handlers remove bot handlers to bot.routers package
